feat(zlib): support for zstd compression/decompression
